Refactored the latest scripts
