+++ tests/functional/features/pr_includes_a_file_which_is_not_chart_related.feature
@@ -7,7 +7,7 @@ Feature: PR includes a non chart related file
 
   Scenario Outline: A user submits a chart with non chart related file
     Given the vendor <vendor> has a valid identity as <vendor_type>
-    And an error-free chart source is used in <chart_path>
+    And chart source is used in <chart_path>
     And user adds a non chart related file
     When the user sends a pull request with both chart and non related file
     Then the pull request is not merged
